requirements.txt
# Core dependencies
jupyterlab>=4.4.3
jupyterlab-git>=0.44.0
jupyter-app-launcher>=0.1.0
dotenv~=0.9.9
voila~=0.5.8
OpenAI~=1.97.0
langgraph~=0.5.3
langchain-nvidia-ai-endpoints~=0.3.12
pydantic~=2.11.7
tavily-python~=0.7.10

# FastAPI backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Backend dependencies
PyPDF2>=3.0.0
python-docx>=0.8.11
faiss-cpu>=1.7.4
langchain>=0.1.0
jinja2>=3.1.0
reportlab>=4.0.0
markdown>=3.5.0

